Moving Migration trait to core diesel library
