+++ diesel_cli/Cargo.toml
@@ -37,6 +37,7 @@ lint = ["clippy"]
 postgres = ["diesel/postgres", "infer_schema_internals/postgres", "url"]
 sqlite = ["diesel/sqlite", "infer_schema_internals/sqlite"]
 mysql = ["diesel/mysql", "infer_schema_internals/mysql", "url"]
+barrels = ["migrations_internals/barrel", "barrel"]
 
 [[test]]
 name = "tests"
